Refactored some code to make it easier to find things. Added more detail to the README.
